+++ backend-node/src/di/controllers/feedbackAnswer.js
@@ -1,13 +1,5 @@
 module.exports = {
     services: {
-        'controllers.feedbackAnswers.stats': {
-            class: 'controllers/FeedbackAnswers/StatsController',
-            arguments: [
-                '@repositories.feedbackQuestion',
-                '@repositories.feedbackAnswer'
-            ]
-        },
-
         'controllers.feedbackAnswers.store': {
             class: 'controllers/FeedbackAnswers/StoreController',
             arguments: [
