+++ backend-node/src/di/controllers/feedbackQuestion.js
@@ -3,6 +3,14 @@ module.exports = {
         'controllers.feedbackQuestions.index': {
             class: 'controllers/FeedbackQuestions/IndexController',
             arguments: ['@repositories.feedbackQuestion']
+        },
+
+        'controllers.feedbackQuestions.stats': {
+            class: 'controllers/FeedbackQuestions/StatsController',
+            arguments: [
+                '@repositories.feedbackQuestion',
+                '@repositories.feedbackAnswer'
+            ]
         }
     }
 };
