+++ backend-node/src/routes/feedback-answers.js
@@ -7,14 +7,11 @@ const validate = require('../middlewares/validate');
 const feedbackAnswerValidator = require('../validators/feedback-answer');
 
 const {
-    Role: { ADMIN, MANAGER, EMPLOYEE }
+    Role: { MANAGER, EMPLOYEE }
 } = require('../models');
 
 module.exports = di => {
-    const statsController = di.get('controllers.feedbackAnswers.stats');
     const storeController = di.get('controllers.feedbackAnswers.store');
-
-    router.get('/stats', loggedOnly(ADMIN), invoke(statsController));
 
     router.post(
         '/',
