+++ backend-node/src/routes/feedback-questions.js
@@ -5,13 +5,15 @@ const invoke = require('../middlewares/invoke');
 const loggedOnly = require('../middlewares/loggedOnly');
 
 const {
-    Role: { MANAGER, EMPLOYEE }
+    Role: { ADMIN, MANAGER, EMPLOYEE }
 } = require('../models');
 
 module.exports = di => {
     const indexController = di.get('controllers.feedbackQuestions.index');
+    const statsController = di.get('controllers.feedbackQuestions.stats');
 
     router.get('/', loggedOnly(MANAGER, EMPLOYEE), invoke(indexController));
+    router.get('/stats', loggedOnly(ADMIN), invoke(statsController));
 
     return router;
 };
